refactor: rename internal files
