+++ packages/ngx-fast-lib/src/index.ts
@@ -1,11 +1,11 @@
 // tokens
-export * from './lib/token/icon-options.model';
-export * from './lib/token/icon-options.token';
-export * from './lib/token/icon-load.strategy.model';
-export * from './lib/token/icon-load.strategy';
+export * from './lib/token/svg-options.model';
+export * from './lib/token/svg-options.token';
+export * from './lib/token/svg-load.strategy.model';
+export * from './lib/token/svg-load.strategy';
 // service
-export * from './lib/icon-registry.service';
+export * from './lib/svg-registry.service';
 // component
-export * from './lib/fast-icon.component';
+export * from './lib/fast-svg.component';
 // module
-export * from './lib/fast-icon.module';
+export * from './lib/fast-svg.module';
