+++ packages/ngx-fast-lib/src/lib/fast-svg.module.ts
@@ -1,12 +1,12 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
-import { FastSvgComponent } from './fast-icon.component';
-import { SvgRegistry } from './icon-registry.service';
+import { FastSvgComponent } from './fast-svg.component';
+import { SvgRegistry } from './svg-registry.service';
 import { CommonModule } from '@angular/common';
 import { FastSvgProviderOptions } from './provider-config.interface';
-import { SvgLoadStrategyImpl } from './token/icon-load.strategy';
-import { SvgLoadStrategy } from './token/icon-load.strategy.model';
-import { SvgOptionsToken } from './token/icon-options.token';
-import { SvgOptions } from './token/icon-options.model';
+import { SvgLoadStrategyImpl } from './token/svg-load.strategy';
+import { SvgLoadStrategy } from './token/svg-load.strategy.model';
+import { SvgOptionsToken } from './token/svg-options.token';
+import { SvgOptions } from './token/svg-options.model';
 
 @NgModule({
   imports: [CommonModule],
