+++ packages/ngx-fast-lib/src/lib/token/svg-load.strategy.ts
