+++ packages/ngx-fast-lib/src/lib/token/svg-options.model.ts
