+++ packages/ngx-fast-lib/src/lib/token/svg-options.token.ts
@@ -1,5 +1,5 @@
 import { InjectionToken } from '@angular/core';
-import { SvgOptions } from './icon-options.model';
+import { SvgOptions } from './svg-options.model';
 
 export const SvgOptionsToken = new InjectionToken<SvgOptions>(
   '__SvgOptionsToken__'
